LayoutTests/platform/wk2/http/tests/security/mixedContent/insecure-executable-css-with-secure-cookies-expected.txt

main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
CONSOLE MESSAGE: line 6: secureCookie=yes
CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-executable-css-with-secure-cookies.html was not allowed to run insecure content from http://127.0.0.1:8080/security/resources/insecure.css.

main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This test opens a window that loads a secure style sheet with insecure content after reading secure cookies. This should block loading of the insecure content since secure cookie was read.
